requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
aiofiles>=0.8.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
openai>=0.27.0
tiktoken>=0.3.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
aiosqlite>=0.17.0
python-dotenv>=0.19.0 